third_party/WebKit/LayoutTests/fast/dom/Document/clone-node.html

<!DOCTYPE html>
<link rel="help" href="http://dom.spec.whatwg.org/#dom-node-clonenode">
<script src="../../../resources/js-test.js"></script>
<script>
description('Tests cloneNode for Document.');

function className(object) {
    return Object.prototype.toString.call(object).slice(8, -1);
}

var doc = document.implementation.createDocument('', 'root', null);

shouldBe('doc.cloneNode(false).__proto__', 'XMLDocument.prototype');
shouldBeEqualToString('className(doc.cloneNode(false))', 'XMLDocument');
shouldBeEqualToString('doc.cloneNode(true).documentElement.localName', 'root');
shouldBeEqualToString('document.cloneNode(true).compatMode', 'CSS1Compat');
shouldBe('document.cloneNode(false).URL', 'document.URL');
shouldBe('document.cloneNode(false).baseURI', 'document.baseURI');
shouldBe('document.cloneNode(false).characterSet', 'document.characterSet');

</script>
</body>
